Updates following code review
