+++ app/routes/interviews.js
@@ -1,7 +1,8 @@
+const Pagination = require('../data/pagination-utils')
 const utils = require('../data/application-utils')
-const { v4: uuidv4 } = require('uuid')
-const { DateTime } = require('luxon')
 const _ = require('lodash');
+const { DateTime } = require('luxon')
+const { v4: uuidv4 } = require('uuid')
 
 function getTimeObject(time) {
   var hours;
@@ -83,11 +84,11 @@ module.exports = router => {
     interviews = interviews.slice(9, interviews.length)
 
     // Get the pagination data
-    let pagination = utils.getPagination(interviews, req.query.page, req.query.limit)
+    let pagination = Pagination.getPagination(interviews, req.query.page, req.query.limit)
 
     let now = interviews[0].interview.date
 
-    interviews = utils.getDataByPage(interviews, req.query.page, req.query.limit)
+    interviews = Pagination.getDataByPage(interviews, req.query.page, req.query.limit)
     interviews = groupInterviewsByDate(interviews)
 
     res.render('interviews/index', {
@@ -100,8 +101,8 @@ module.exports = router => {
   router.get('/interviews/past', (req, res) => {
     let interviews = getInterviews(req.session.data.applications)
     interviews = interviews.slice(0, 9).reverse()
-    let pagination = utils.getPagination(interviews, req.query.page, req.query.limit)
-    interviews = utils.getDataByPage(interviews, req.query.page, req.query.limit)
+    let pagination = Pagination.getPagination(interviews, req.query.page, req.query.limit)
+    interviews = Pagination.getDataByPage(interviews, req.query.page, req.query.limit)
     interviews = groupInterviewsByDate(interviews)
     res.render('interviews/past', {
       interviews,
